Work in Progress
es ist jetzt möglich items anzulegen

+++ Kühlschrankplaner/0-plugins/src/main/java/com/Kuhlschrankmanufaktur/Kuhlschrankplaner/persistance/DataKühlschrankRepository.java
@@ -3,5 +3,5 @@ package com.Kuhlschrankmanufaktur.Kuhlschrankplaner.persistance;
 import com.Kuhlschrankmanufaktur.Kuhlschrankplaner.domain.Kühlschrank;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DataKühlschrankRepository extends JpaRepository<Kühlschrank, Long> {
+public interface DataKühlschrankRepository extends JpaRepository<Kühlschrank, Integer> {
 }
